README.md
After nearly a year, I have begun working on this project again. The goal is to create a website that can be used by future engineering students to help learn linear algebra in a more concise and easy-to-understand manner compared to what is commonly presented in most textbooks. This is reflected by the emphasis on intuition leading to short but powerful proofs of mathematical theorems; visual aids and animations to help conceptualize abstract concepts; and interactive software to increase the engagement of the material being presented. 

Please note that this is not intended for mathematics or computer science students who may desire a more rigorous approach to the material being presented here.